agent/pyproject.toml
[project]
name = "deep-fractional-agent"
version = "0.1.0"
description = "LangChain Deep Agents backend for Fractional Quest"
requires-python = ">=3.12"
dependencies = [
    "copilotkit>=0.1.76",
    "deepagents>=0.3.5",
    "ag-ui-langgraph>=0.0.23",
    "fastapi>=0.115.14",
    "langchain>=1.2.4",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "langgraph-checkpoint>=4.0.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
    "asyncpg>=0.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
